+++ src/modules/default/index.js
@@ -3,8 +3,7 @@ import * as meteor from '../meteor/';
 import * as mongo from '../mongo/';
 
 import debug from 'debug';
-import path from 'path';
-import { resolve } from '../utils';
+import { resolvePath } from '../utils';
 import sh from 'shelljs';
 
 const log = debug('mup:module:default');
@@ -16,17 +15,17 @@ export function deploy(api) {
   return meteor.deploy(api);
 }
 
-export function help(api) {
+export function help() {
   log('exec => mup help');
 }
-export function init(/* api */) {
+export function init() {
   log('exec => mup init');
 
   // TODO check if mup.js or settings.json files exists
-  const mupJs = resolve(__dirname, 'template/mup.js.sample');
-  const settinsJson = resolve(__dirname, 'template/settings.json');
-  const mupJsDst = resolve(process.cwd(), 'mup.js');
-  const settingsJsonDst = resolve(process.cwd(), 'settings.json');
+  const mupJs = resolvePath(__dirname, 'template/mup.js.sample');
+  const settinsJson = resolvePath(__dirname, 'template/settings.json');
+  const mupJsDst = resolvePath(process.cwd(), 'mup.js');
+  const settingsJsonDst = resolvePath(process.cwd(), 'settings.json');
 
   sh.cp(mupJs, mupJsDst);
   sh.cp(settinsJson, settingsJsonDst);
@@ -39,30 +38,26 @@ export function logs(api) {
 
 export function reconfig(api) {
   log('exec => mup reconfig');
-  return meteor.envconfig(api)
-    .then(() => meteor.start(api));
+  return meteor.envconfig(api).then(() => meteor.start(api));
 }
 
 export function restart(api) {
   log('exec => mup restart');
-  return meteor.stop(api)
-    .then(() => meteor.start(api));
+  return meteor.stop(api).then(() => meteor.start(api));
 }
 
 export function setup(api) {
   log('exec => mup setup');
   const config = api.getConfig();
-  return docker.setup(api)
-    .then(() => {
-      if (config.mongo) {
-        return Promise.all([
-          meteor.setup(api),
-          mongo.setup(api)
-        ])
-          .then(() => (mongo.start(api)));
-      }
-      return meteor.setup(api);
-    });
+  return docker.setup(api).then(() => {
+    if (config.mongo) {
+      return Promise.all([
+        meteor.setup(api),
+        mongo.setup(api)
+      ]).then(() => mongo.start(api));
+    }
+    return meteor.setup(api);
+  });
 }
 
 export function start(api) {