+++ src/modules/docker/index.js
@@ -1,11 +1,10 @@
 import debug from 'debug';
 import nodemiral from 'nodemiral';
-import path from 'path';
-import { resolve } from '../utils';
-import {runTaskList} from '../utils';
+import { resolvePath } from '../utils';
+import { runTaskList } from '../utils';
 const log = debug('mup:module:docker');
 
-export function help(/* api */) {
+export function help() {
   log('exec => mup docker help');
 }
 
@@ -14,15 +13,18 @@ export function setup(api) {
   const list = nodemiral.taskList('Setup Docker');
 
   list.executeScript('setup docker', {
-    script: resolve(__dirname, 'assets/docker-setup.sh')
+    script: resolvePath(__dirname, 'assets/docker-setup.sh')
   });
 
-  const sessions = api.getSessions([ 'meteor', 'mongo', 'proxy' ]);
-  const rsessions = sessions.reduce((prev, curr) => {
-    if (prev.map(session => session._host).indexOf(curr._host) === -1) {
-      prev.push(curr);
-    }
-    return prev;
-  }, []);
+  const sessions = api.getSessions(['meteor', 'mongo', 'proxy']);
+  const rsessions = sessions.reduce(
+    (prev, curr) => {
+      if (prev.map(session => session._host).indexOf(curr._host) === -1) {
+        prev.push(curr);
+      }
+      return prev;
+    },
+    []
+  );
   return runTaskList(list, rsessions);
 }