+++ src/validate/servers.js
@@ -2,15 +2,15 @@ import { combineErrorDetails } from './utils';
 import joi from 'joi';
 
 const schema = joi.object().keys({
-  host: joi.alternatives(
-    joi.string().ip({
-      version: [
-        'ipv4',
-        'ipv6'
-      ]
-    }),
-    joi.string().uri()
-  ).required(),
+  host: joi
+    .alternatives(
+      joi.string().ip({
+        version: ['ipv4', 'ipv6']
+      }),
+      joi.string().uri(),
+      joi.string().trim()
+    )
+    .required(),
   username: joi.string().required(),
   pem: joi.string().trim(),
   password: joi.string(),
@@ -21,9 +21,9 @@ const schema = joi.object().keys({
 
 export default function validateServers(servers) {
   let details = [];
-  Object.keys(servers).forEach((key) => {
+  Object.keys(servers).forEach(key => {
     let result = joi.validate(servers[key], schema, {
-        convert: false
+      convert: false
     });
     details = combineErrorDetails(details, result);
   });