+++ CHANGELOG.md
@@ -1,10 +1,17 @@
+# 1.2.2
+- Configure nginx max client upload size, and increase default to `10M` (@shadowcodex)
+- Displays better message if it can not find the meteor app
+- Displays message if can not find pem for server
+- Improve validating server's `host` in config
+- Validator checks for `http://` or `https://` in `ROOT_URL`
+- Update documentation on using `mup` on Windows
+
 # 1.2.1
 - All paths support "~"
 - Add `server` and `allowIncompatibleUpdates` to build config (@alvelig)
 - Allow `mobile-settings` build option to use `settings.json` file (@alvelig)
 - Add `mup --version` command
 - Fix validating env `variables` and `imageFrontendServer`
-
 
 # 1.2.0
 - Support Meteor 1.4 by default (@ffxsam)